Generate concurrency-safe code in test bundle (#1274)
With the new concurrency features in Swift 6, there are also a set of
checks that ensure that the code is safe to run concurrently. Public
static mutable variables are considered unsafe unless they are
specifically of a type that conforms to `Sendable`. This patch changes
the test bundle generation to use `let` instead of `var` in the case of
`public static` properties.