Check that we don't cause an out of bounds crash when accessing the new values.
Added an assertion that checks that the cursor plus the lenght of the memory
to be decoded is less or equal than the length of the entire memory for the
raw value that has been passed, avoiding a runtime crash.

Also both if statements were combined into a signgle one since, as pointed out
in [1], `isBeta` and `isExternal` have to either both exists or not.

---

[1] https://github.com/swiftlang/swift-docc/pull/1249#discussion_r2259784235