Adds a shared utility for determining beta status
Adds an extension to `Array<AvailabilityRenderItem>` which defines a property `isBeta` with the same logic currently defined in 3 separate locations:
- `NavigatorIndexableRenderMetadataRepresentation` [1]
- `OutOfProcessReferenceResolver.ResolvedInformation` [2]
- `LinkDestinationSummary` [3]

This will eliminate the possibility that these 3 implementations will inadvertently become out of sync with each other. The logic remains that an item is in beta if all platforms are marked as beta.

There is still at least one more place in the code-base which determines beta status [4], but that remains hard to merge at the moment.

[1]: https://github.com/swiftlang/swift-docc/blob/1c54e106f5c9a92a69225492b1a89c934677ae58/Sources/SwiftDocC/Indexing/Navigator/RenderNode%2BNavigatorIndex.swift#L126-L134
[2]: https://github.com/swiftlang/swift-docc/blob/1c54e106f5c9a92a69225492b1a89c934677ae58/Sources/SwiftDocC/Infrastructure/External%20Data/OutOfProcessReferenceResolver.swift#L592-L598
[3]: https://github.com/swiftlang/swift-docc/blob/1c54e106f5c9a92a69225492b1a89c934677ae58/Sources/SwiftDocC/Infrastructure/Link%20Resolution/ExternalPathHierarchyResolver.swift#L186-L192
[4]: https://github.com/swiftlang/swift-docc/blob/1c54e106f5c9a92a69225492b1a89c934677ae58/Sources/SwiftDocC/Model/Rendering/DocumentationContentRenderer.swift#L210