Populate symbol kind in external render nodes (#1251)
* Add mapping from documentation kind to symbol kind

There is a mapping from `SymbolGraph.Symbol.KindIdentifier` to `DocumentationNode.Kind`, but not the other way around; this commit adds one.

For documentation kinds which both map to the same symbol kind, this has been expressed as:
- both documentation kinds are converted to the same symbol kind in `symbolKind(for:)`.
- a specific documentation kind is chosen as the default mapping in `kind(forKind:)`

For example, for `DocumentationNode.Kind. typeConstant` [1]:
- both `symbolKind(for: .typeConstant)` and `symbolKind(for: .typeProperty)` return `.typeProperty`
- `kind(forKind: .typeProperty)` returns `.typeProperty`

This function will be used to map and external entity's documentation kind to a symbol kind, so that that information can be populated as part of the navigator.

## Verification:

This has been verified by testing the round trip conversion of all symbol kinds, and all documentation kinds which are symbols.

[1]: https://github.com/swiftlang/swift-docc/blob/65aaf926ec079ddbd40f29540d4180a70af99e5e/Sources/SwiftDocC/Model/Kind.swift#L151

* Capture symbol kind as part of external entity

Adds a new property to `LinkResolver.ExternalEntity` which stores the symbol kind of the external entity.

This is derived at the level of `OutOfProcessReferenceResolver.ResolvedInformation`, which has access to the documentation kind [1] and already uses it to derive the render node kind and role, but then discards it.

This commit adds logic to `OutOfProcessReferenceResolver.ResolvedInformation` to derive the symbol kind from the documentation kind, then capture that in `LinkResolver.ExternalEntity.

We need access to the documentation kind in order to resolve the symbol kind as part of computing the navigator title. [2]

[1]: https://github.com/swiftlang/swift-docc/blob/65aaf926ec079ddbd40f29540d4180a70af99e5e/Sources/SwiftDocC/Infrastructure/External%20Data/OutOfProcessReferenceResolver.swift#L564-L565
[2]: https://github.com/swiftlang/swift-docc/blob/65aaf926ec079ddbd40f29540d4180a70af99e5e/Sources/SwiftDocC/Indexing/Navigator/RenderNode%2BNavigatorIndex.swift#L163

* Propagate symbol kind to navigator for external nodes

Propagates the symbol kind value from `LinkResolver.ExternalEntity` to the `ExternalRenderNode` used for generating the navigation index for external render nodes.
This completes adding symbol kind support for external entities in the navigator.

Also updates how the symbol kind is propagated to the navigator metadata, by using `.renderingIdentifier` over `.identifier`, to match the behaviour of local nodes [1].

Fixes rdar://156487928.

[1]: https://github.com/swiftlang/swift-docc/blob/65aaf926ec079ddbd40f29540d4180a70af99e5e/Sources/SwiftDocC/Model/Rendering/RenderNodeTranslator.swift#L1300