Fix `@SupportedLanguage` directive for articles
The `@SupportedLanguage` directive allows specifying a language that an
article is available in. It can be used within the `@Metadata` directive
like in the below example, where the article is made available in both
Swift and Objective-C:

```

@Metadata {
  @SupportedLanguage(swift)
  @SupportedLanguage(objc)
}
```

This directive is processed when creating the topic graph node for the
article. In cases where the article is the root page or a standalone
article (when no technology root is defined), the directive is correctly
processed as a part of registering the article in the documentation
cache. However, in the case of a non-root article in a catalog
containing a root article, this step was being skipped. This patch fixes
this to ensure that the directive is processed for all articles in a
catalog.

rdar://160284853