Speed up directory monitoring tests
The directory monitoring tests check for whether the documentation is
rebuilt if the documentation catalog is edited when running a live
server with `docc preview`. These tests assumed an extremely generous
timeout for receiving the directory change signal, and ran much slower
than the rest of the test suite. The timeouts have been reduced to a
more reasonable threshold that succeeds even with throttled resources on
a single-core machine.