Store all source languages in documentation node (#1326)
The documentation node stores the list of source languages that the node
is available in. Previously, this information was being stored in the
render node as a set containing the singular source language of the
node. This worked in the past when Swift was the only relevant language,
but `sourceLanguage` is now relegated to a backwards-compatible property
that queries `sourceLanguages` for the value it returns. This patch
updates the documentation node constructors to correctly store the set
of source languages in the node.

rdar://163034867