Emit variants if any root module is a symbol
When the render node translator visits articles, it emits variants for
each language that the article is supported in. This step is skipped if
an article-only catalog is being built, since it would otherwise lead to
a redundant "Swift" language marker being rendered in the page. However,
if a catalog contains more than one module, i.e. an article-only
collection and a different module's symbol graph in the same catalog,
then the variants are not emitted despite the possibility of the page
being available in more than one language. This commit changes the
boolean check to skip emitting variants only if all root modules are
article-only. It does not affect the behaviour when building a correctly
configured catalog that contains a sole root module.

rdar://163926698