Exclude CMakeLists.txt files from SwiftPM targets
SwiftPM requires that all files within a target are accounted for as
part of the build. If there are extraneous files that do not form a part
of the target, the package manager emits a warning about unhandled
files. This patch updates Package.swift to exclude all CMakeLists.txt
files from their relevant SwiftPM targets to remove this warning.