Encode absolute presentation URL if present
Support for absolute URLs in link summaries was added in 7db36ad21303f2b85c094b7add3023b4c9e1111e.

It added support for absolute URLs when decoding a link summary, but not when encoding it:
https://github.com/swiftlang/swift-docc/blob/cfcd96f0e992af2287661f6a26e8398096f6bc1e/Sources/SwiftDocC/LinkTargets/LinkDestinationSummary.swift#L772-L773
https://github.com/swiftlang/swift-docc/blob/cfcd96f0e992af2287661f6a26e8398096f6bc1e/Sources/SwiftDocC/LinkTargets/LinkDestinationSummary.swift#L728

Both are needed when using an external link resolver, as the external link resolver needs to encode the entity to send to DocC.

Fixes rdar://164628218.