RenderBlockContent.Aside Names (#1361)
* Introduce a new stored property on `RenderBlockContent.Aside` called
`name` to stored the name or title of aside blocks. This allows DocC to
independently save and control the styling and the name of asides.
Previously DocC conflated the style of the aside block (e.g. colors or
other styling) with the name of the aside. Although it isn't currently
possible to author an aside with a different style and name, this change
allows the `Aside` model itself to support that one day.

Aside now has four available initializers:
* init(style: AsideStyle, content: [RenderBlockContent])
* init(name: String, content: [RenderBlockContent])
* init(style: AsideStyle, name: String, content: [RenderBlockContent])
* init(asideKind: Markdown.Aside.Kind, content: [RenderBlockContent])

Most often DocC will use the fourth initializer to create aside blocks
from markdown, via RenderContentCompiler. This initializer, creating an
aside from a Swift Markdown Aside.Kind, also contains special logic to
determine a capitalized name for the aside that was previously
implemented in AsideStyle.

Simplify the `RenderBlockContent.AsideStyle` structure to act as if it
were an enum of the known styles supported by DocC Render: note, tip,
experiment, important, or warning. (Since this is public API, this
change retains AsideStyle as a structure with a string raw value.)

Also simplify the JSON encoding and decoding of Aside and AsideStyle
structures to save and read the style and name in a natural way.

* Correct Swift syntax error that cropped up in macOS CI

* Improve the documentation comments for the RenderBlockContent.Aside and
AsideStyle initializers.

* Update Sources/SwiftDocC/Model/Rendering/RenderContentCompiler.swift

Co-authored-by: David Rönnqvist <ronnqvist@apple.com>

* Simplify RenderBlockContent.AsideStyle.init(rawValue:)

* Update Sources/SwiftDocC/Model/Rendering/Content/RenderBlockContent.swift

Co-authored-by: David Rönnqvist <ronnqvist@apple.com>

* When decoding JSON, AsideStyle#init(from:) calls init(rawValue:) to
handle uppercased and other invalid style values.

Fix the logic for decoding RenderBlockContent asides, to retain both the
name and style separately when they are different in JSON. This is
different than the previous behavior, which overrode the style with the
name.

* Cleanup: Update copywrite dates, remove stray debug print statements and
whitespace, and simplify how we express some expected data in a unit
test.

* Deprecate RenderBlockContent.AsideStyle.displayName

* Simplify some test code in SemaToRenderNodeTests.swift and RenderBlockContent+AsideTests.swift

* Refactor RenderBlockContent_AsideTests.decodeAsideRenderBlock to return
a non-optional Aside by implementing the same logic the #require uses. And remove #require from all the call sites.

* Remove use of deprecated AsideStyle.displayName at one call site.

* Update Tests/SwiftDocCTests/Rendering/RenderBlockContent+AsideTests.swift

Co-authored-by: David Rönnqvist <ronnqvist@apple.com>

* Refactor RenderBlockContent_AsideTests.decodeAsideRenderBlock again to
avoid throwing an exception manually, and to record the source location
when decoding an unexpected block type.

---------

Co-authored-by: David Rönnqvist <ronnqvist@apple.com>