Support reading "extended" module from symbol graph instead of the file's name.
This is a companion change to the change in the Swift compiler to emit the "extended" module as an additional field in the symbol graph. Reading the base and extended modules from the file name is not feasible for module's with long names. A new compiler option was added in [Swift #83782](https://github.com/swiftlang/swift/pull/83782) to allow emitting symbol graph file's without both module names in the file name. The shortened format isn't yet supported by swift-docc though.

This also relies on a corresponding change to swift-docc-symbolkit where `GraphCollector.moduleNameFor` is updated.

The change here consolidates some `moduleNameFor` functionality where swift-docc was implementing similar behavior as exists in swift-docc-symbolkit, while already depending on symbolkit so it seems prudent to just share the implementation of `moduleNameFor`.