Clean up fallback platform deconstruction
The `DefaultAvailability.fallbackPlatforms` variable provides a mapping
from platforms to their respective fallbacks. This map was being
deconstructed with incorrect terminology in a few places, with the
platform itself being called "fallback", and the fallback platform being
called "inherited". This has been cleaned up to make the nomenclature
consistent across the codebase.