Include fallback platforms in articles
When DocC processes the availability information for symbol graphs, it
identifies if fallback platforms are available, and if so, injects
platforms that inherit their availability from the fallback platforms.
E.g. the iPadOS and Mac Catalyst platforms inherit their availability
from iOS.

In the case of articles, the platform availability information is not
used in rendering the page. However, sample code pages are processed as
articles with a different page kind, and these pages do render the
platform availability information. Hence, it is necessary to also
process the fallback platforms when setting the availability for an
article's render node in order to correctly display the list of
supported platforms for sample code pages. This patch updates the logic
in the render node translator to process the availability information
for articles correctly and include fallback platforms.

rdar://158230351