[6.3] Include availability for fallback platforms in articles (#1410)
* Clean up fallback platform deconstruction

The `DefaultAvailability.fallbackPlatforms` variable provides a mapping
from platforms to their respective fallbacks. This map was being
deconstructed with incorrect terminology in a few places, with the
platform itself being called "fallback", and the fallback platform being
called "inherited". This has been cleaned up to make the nomenclature
consistent across the codebase.

* Include fallback platforms in articles

When DocC processes the availability information for symbol graphs, it
identifies if fallback platforms are available, and if so, injects
platforms that inherit their availability from the fallback platforms.
E.g. the iPadOS and Mac Catalyst platforms inherit their availability
from iOS.

In the case of articles, the platform availability information is not
used in rendering the page. However, sample code pages are processed as
articles with a different page kind, and these pages do render the
platform availability information. Hence, it is necessary to also
process the fallback platforms when setting the availability for an
article's render node in order to correctly display the list of
supported platforms for sample code pages. This patch updates the logic
in the render node translator to process the availability information
for articles correctly and include fallback platforms.

rdar://158230351